requirements.txt
asgiref==3.11.1
Django==4.2.11
djangorestframework==3.15.2
drf-yasg==1.21.7
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
reportlab==4.4.9